Moved the generateUri function into the main index file and imported into the test file.
